Big change.
Due to not being able to get the parts for the full system, we’re now just laying out the slots. We can add in the video and ethernet as cards rather than them being provided as-is.

This simplifies things a lot, now we only have 1 STM chip, and its associated SDRAM. The board is a lot smaller as well, and fits into 10cm in the Y dimension, which means making cards can be done using the cheap 10cm x 10cm board sizes that places like Seeed offer.

Need to define a cartridge port symbol, and remap all the pins to the STM32F746BGT6 (208 pin LQFP) which is available in China.

Assumption here is that the host-side will have an XPLA3 CPLD sitting on it, which manages the signal directionality under control of the STM, and also acts as a signal-buffer/level-translator. All signals on the expansion box are 3.3v apart from the 5v supply.

In the reverse direction, the STM ought to have sufficient juice to send the signals back on its own, since there’s only 1 load, and the signal path is short.